% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_setup_exercise.R
\name{rt_add_task}
\alias{rt_add_task}
\title{Add task from database to script}
\usage{
rt_add_task(tdb, id, task_nr = id, script_nr = 1, dir = ".", folder = "Quiz1")
}
\arguments{
\item{tdb}{Task database (data.frame with the columns ID, Task, Test)}

\item{id}{ID to be selected from \code{tdb}, e.g. "Q10_2"}

\item{task_nr}{Task number to be printed in \code{script_x.R}, DEFAULT: id}

\item{script_nr}{Script number to be used for this task. DEFAULT: 1}

\item{dir}{Directory, pasted together with \code{folder}. DEFAULT: "."}

\item{folder}{Folder at \code{dir}. DEFAULT: "Quiz1"}
}
\value{
dir
}
\description{
Add a task from the database to \code{script_x.R}
and the corresponding tests to \code{tests.R}.
}
\seealso{
\link{rt_setup_exercise}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
