% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_check_for_unsaved_files.R
\name{rt_check_for_unsaved_files}
\alias{rt_check_for_unsaved_files}
\title{Check for files with unsaved changes}
\usage{
rt_check_for_unsaved_files(dir = ".", warnonly = FALSE)
}
\arguments{
\item{dir}{Path to (processed) exercise folder.
Must contain ".Rproj.user" folder. DEFAULT: "."}

\item{warnonly}{Logical: Only issue a \link{warning} instead of an
error with \link{stop}? DEFAULT: FALSE}
}
\value{
NULL if all files are saved. Otherwise, if \code{warnonly=TRUE}, the names of changed but unsaved files.
}
\description{
Raise an informative error/warning message if any of the opened
Rstudio R project file tabs contain unsaved changes.
}
\seealso{
\link{rt_score}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
