% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_contains.R
\name{rt_contains}
\alias{rt_contains}
\title{test object value}
\usage{
rt_contains(obj, value, fixed = TRUE, ignore_space = TRUE, ignore_quote = TRUE)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{value}{value \code{obj} should have, can be char / numeric / other.}

\item{fixed}{Fixed match in \code{\link[=grepl]{grepl()}}? DEFAULT: TRUE}

\item{ignore_space}{Remove spaces before comparison? DEFAULT: TRUE}

\item{ignore_quote}{Replace \verb{'} with \verb{"} before comparison? DEFAULT: TRUE}
}
\value{
Logical: TRUE / FALSE depending on whether object contains \code{value}.
}
\description{
Test whether object contains certain value(s), calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\link{rt_has_value}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
