% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_create.R
\name{rt_create}
\alias{rt_create}
\title{Create exercise folder from zip file}
\usage{
rt_create(zipfile = NULL, deletezip = TRUE, ...)
}
\arguments{
\item{zipfile}{Path to zip file (or, if unzipped, any of the files within the folder).
"\" (on Windows) must be changed to "/".
DEFAULT: NULL, meaning to use \code{\link[=file.choose]{file.choose()}}}

\item{deletezip}{If the exercise folder was created sucessfully, delete the original zip file? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{unzip}}}
}
\value{
exdir, invisibly
}
\description{
Create exercise folder from zip file, with a \code{.Rproj} file and try to open that in Rstudio.
On Mac OS, the exercise file gets unzipped upon downloading, hence any file within the folder can be used.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct-Dec 2020
}
\keyword{file}
