% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_create_task.R
\name{rt_create_task}
\alias{rt_create_task}
\title{Create Task folder from zip file}
\usage{
rt_create_task(zipfile, exdir = NULL, ...)
}
\arguments{
\item{zipfile}{Path to zip file. Remember: on Windows, "\" must be changed to "/".}

\item{exdir}{Folder to unzip to, e.g. "." for current wd.
DEFAULT: NULL (path from zipfile).}

\item{\dots}{Further arguments passed to \code{\link{unzip}}}
}
\value{
ReturnValue
}
\description{
Create Task folder from zip file, with a \code{.Rproj} file and try to open that in Rstudio.
}
\examples{
# rt_create_task("C:/Dropbox/R/FREELANCING/HPI_2020_R/CodeOceanTesting/FProg20_R_quiz_1.zip")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
