% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_due_warn.R
\name{rt_due_warn}
\alias{rt_due_warn}
\title{Warn about exercise submission time}
\usage{
rt_due_warn(
  due,
  tz = "CET",
  format = "\%Y-\%m-\%d \%H:\%M",
  begin = 90,
  end = -15,
  ...
)
}
\arguments{
\item{due}{Charstring with due time in the format "2020-11-30 16:00".
Remember 15:00 UTC is 16:00 CET.}

\item{tz}{Timezone for \link{strptime}. DEFAULT: "CET"}

\item{format}{Format for \link{strptime}. DEFAULT: "\%Y-\%m-\%d \%H:\%M"}

\item{begin}{Numeric: remaining time in minutes after which the message is displayed. DEFAULT: 90}

\item{end}{Numeric: remaining time in minutes after which the message is no longer displayed. DEFAULT: -15}

\item{\dots}{Further arguments passed to \code{\link[=rt_warn]{rt_warn()}}}
}
\value{
output of \code{\link[=rt_warn]{rt_warn()}}
}
\description{
Warn exercise participant about due time.
}
\examples{
rt_due_warn("2020-11-30 16:00")
rt_due_warn("2020-11-30 16:00", begin=1e9, end=1e9)

rt_due_warn(Sys.time()+125, format="\%F \%T") # format for sub-minute accuracy
rt_due_warn(Sys.time()+30, format="\%F \%T")

}
\seealso{
\code{\link[=rt_warn]{rt_warn()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2020
}
