% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_env.R
\name{rt_env}
\alias{rt_env}
\title{interact with rt_test_env}
\usage{
rt_env(
  id = NULL,
  lang = NULL,
  pass = NULL,
  fail = NULL,
  info = NULL,
  script = NULL
)
}
\arguments{
\item{id}{Number or charstring to be set as task ID for \link{rt_warn} messages.
DEFAULT: NULL}

\item{lang}{Charstring with language for \link{rt_warn} messages.
Currently, "en" and "de" are implemented. DEFAULT: NULL}

\item{pass}{Number of task(s) to set as successfull. DEFAULT: NULL}

\item{fail}{Number of task(s) to set as failed. DEFAULT: NULL}

\item{info}{Free-form information to be added. Used in rt_run_script. DEFAULT: NULL}

\item{script}{Charstring vector to be stored as the most recently run student
script, normally set internally by \link{rt_run_script} so that
\link{rt_test_task} can use it (e.g. for \code{section} or \code{ndefinitions})
without it having to be passed explicitly each time. DEFAULT: NULL}
}
\value{
If \code{id}, \code{lang} or no arguments are given, list with the env content.\cr
If \code{pass} or \code{fail} are given, TRUE or FALSE respectively.\cr
Later arguments are ignored if a prior argument is given.
}
\description{
interact with \code{rt_test_env}, see \link{rt_test_exercise}.\cr
Remember to use \code{\link[=list2env]{list2env()}} on the first call, see \link{rt_warn} examples.
}
\seealso{
\link{rt_test_exercise}, which defines the env at the start
and \link{rt_warn} with some examples
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021, Jan 2022
}
