% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_examine_file.R
\name{rt_examine_file}
\alias{rt_examine_file}
\title{Examine student-written file}
\usage{
rt_examine_file(
  fn,
  sep = NULL,
  quote = NULL,
  dec = ".",
  saf = FALSE,
  header = TRUE,
  rnames = NULL,
  value = NULL,
  stepnames = TRUE,
  ...
)
}
\arguments{
\item{fn}{Name of file to be checked.}

\item{sep}{Column separator that should have been used by students in
\link{write.table}. DEFAULT: NULL (not specifically checked,
any whitespace should be fine)}

\item{quote}{T/F: should file contain quotes? DEFAULT: NULL (not checked)}

\item{dec}{Decimal separator passed to \link{read.table}. DEFAULT: "." (not included in call)}

\item{saf}{\code{stringsAsFactors} passed to \link{read.table}. DEFAULT: FALSE}

\item{header}{\code{header} passed to \link{read.table}. DEFAULT: TRUE}

\item{rnames}{T/F: should row.names have been used in \link{write.table}?
DEFAULT: NULL (not checked)}

\item{value}{Object that should come up with \code{read.table(fn,header,sep,dec,saf)}.
Note: convert factors with \link{as.character} (unless saf=TRUE).
DEFAULT: NULL (not checked)}

\item{stepnames}{Passed to \link{rt_test_object}. DEFAULT: TRUE (unlike normally)}

\item{\dots}{Further arguments passed to \link{rt_test_object}.}
}
\value{
TRUE / FALSE, with \link{rt_warn} message.
}
\description{
Test files written by students.
}
\examples{
\dontrun{ # exclude file writing from CRAN checks
write.table(iris, "iri.txt")
iris2 <- iris ; iris2$Species <- as.character(iris2$Species)
rt_examine_file("iri.txt", sep="\t", value=iris2)
unlink("iri.txt")
}
}
\seealso{
\link{rt_test_task}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2021
}
