% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_exercise_to_db.R
\name{rt_exercise_to_db}
\alias{rt_exercise_to_db}
\title{Format exercise tasks for database}
\usage{
rt_exercise_to_db(dir = "exercise99", ntasks = 10)
}
\arguments{
\item{dir}{Directory with \code{script_1.R}, \code{script_1.R}, \code{...}, \code{tests.R}. DEFAULT: "exercise99"}

\item{ntasks}{Number of tasks to obtain. DEFAULT: 10}
}
\value{
location of tempfile with data.frame with tasks and reduced tests.
}
\description{
Take an entire exercise (\code{script_1.R}, \code{script_1.R}, \code{...}, \code{tests.R})
and format them to be added to the task database at
\url{https://docs.google.com/spreadsheets/d/1ggSOYQ_veXgPmvA8cHCLnASkLUvc6-3t6UhJWPRPVoQ}.\cr
Basically, the opposite of \code{\link[=rt_setup_exercise]{rt_setup_exercise()}}.
}
\examples{
if(FALSE) # suppress automated runs
rt_exercise_to_db("C:/Users/berry/Desktop/CORQUIZ")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2020
}
