% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_exists.R
\name{rt_exists}
\alias{rt_exists}
\title{test object existence}
\usage{
rt_exists(obj)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}
}
\value{
Logical: TRUE / FALSE depending on whether object exists.
}
\description{
Test existence of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
