% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_file2openedlist.R
\name{rt_file2openedlist}
\alias{rt_file2openedlist}
\title{Add files to list of Rstudio opened source documents}
\usage{
rt_file2openedlist(file, dir = ".")
}
\arguments{
\item{file}{Filename to be added.}

\item{dir}{Target directory with a .Rproj file. DEFAULT: "." (getwd())}
}
\value{
expanded filename, invisibly.
}
\description{
Add files to the list of Rstudio opened source documents.
The related .Rproj Rstudio project should be closed at the time of writing.
}
\examples{
# See rt_create_task

}
\seealso{
\code{\link[=rt_create_task]{rt_create_task()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{internal}
