% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_full_file.R
\name{rt_full_file}
\alias{rt_full_file}
\title{Get full file locally}
\usage{
rt_full_file(..., size = 10000)
}
\arguments{
\item{\dots}{File name(s)}

\item{size}{\link{file.size} beyond which files are not replaced. DEFAULT: 1e4}
}
\value{
filename if replacement happened, FALSE with ff-prefixed \link{rt_warn} otherwise
}
\description{
Replace small sample version of a data file with complete one on local machine.\cr
Scoring is relatively slow for large files on CodeOcean. Have a small version there.
With this function, update to a large version locally for smooth graphics.
}
\seealso{
\code{\link{help}}, \code{graphics::\link[graphics]{plot}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2022
}
\keyword{file}
