% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_get_context_id.R
\name{rt_get_context_id}
\alias{rt_get_context_id}
\title{Get contextIdentifier for .Rproj.user ID}
\usage{
rt_get_context_id()
}
\value{
Hexadecimal ID charstring
}
\description{
Internal function to find the User-specific 8 digit hexadecimal id for .Rproj.user ID
}
\seealso{
\code{\link[=rt_file2openedlist]{rt_file2openedlist()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2020
}
\keyword{internal}
