% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_get_echo.R
\name{rt_get_echo}
\alias{rt_get_echo}
\title{get echo from code}
\usage{
rt_get_echo(expr)
}
\arguments{
\item{expr}{Code to be executed while output is collected.
Can be contained within \{curly brackets\}.}
}
\value{
Charstring vector from \code{\link[=readLines]{readLines()}}
}
\description{
get echoed output from running certain code, e.g. messages + \code{\link[=cat]{cat()}} outputs.
}
\examples{

rt_get_echo(cat("things\nwith linebreaks"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2020
}
