% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_gives_warning.R
\name{rt_gives_warning}
\alias{rt_gives_warning}
\title{Test whether an expression yields a specific warning}
\usage{
rt_gives_warning(expr, w = NULL)
}
\arguments{
\item{expr}{Expression to be evaluated}

\item{w}{Charstring: warning that needs to be present (matched with grepl).
Use \code{w=""} to test if no warnings are generated.
Use \code{w=NULL} to test if any warning is produced.
DEFAULT: NULL}
}
\value{
TRUE or FALSE
}
\description{
Test whether an expression yields a specific warning
}
\examples{
rt_test_env <- new.env()
rt_gives_warning({log(-7); as.numeric("a")}, "NaNs produced") # TRUE
rt_gives_warning({log( 7); as.numeric("a")}, "NaNs produced") # wrong warning
rt_gives_warning({log(-7); as.numeric("a")}, "") # if no warning was desired
rt_gives_warning(log(7), "") # TRUE (no warning)
rt_gives_warning(log(-7), "") # shouldn't warn but does
rt_gives_warning(log(-7), NULL) # TRUE (generates any warning)
rt_gives_warning(log(7), NULL) # should warn but doesn't
}
\seealso{
\link{rt_gives_echo} \url{https://stackoverflow.com/a/4947528/1587132}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
