% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_argument.R
\name{rt_has_argument}
\alias{rt_has_argument}
\title{Test function call for certain arguments}
\usage{
rt_has_argument(code, arg, value = NULL, ignore_space = TRUE, notsub = NULL)
}
\arguments{
\item{code}{Charstring with code, normally from \link{rt_select_script_section}}

\item{arg}{Charstring with a single argument name to be tested}

\item{value}{Value the argument should have. Use \code{'escaped "quotation" marks'}.
Can be left NULL if only the presence of the argument is tested, not its value.}

\item{ignore_space}{Remove spaces before comparison? DEFAULT: TRUE}

\item{notsub}{Arguments that should not be treated as a sub-symbol.
E.g. for \code{value="mean(x,na.rm=TRUE)"}, \code{na.rm} in \code{code}
is parsed as a new sub-level and test would fail.
Avoid this manually with \code{notsub="na.rm"}.
Do not set ignore_space to FALSE when using this.
DEFAULT: NULL (ignored)}
}
\value{
Logical: TRUE / FALSE depending on whether condition is met
}
\description{
Test if a function call in stundent's code has certain arguments.
If argument value is a charstring, test with \code{rt_has_argument(code, "arg", '"value"')}.\cr
Use double backslashes as in \code{rt_has_argument(code, "arg", '"value\\\\U{00B0}"')}
for Unicode etc. \cr
Attenton: if the user script contains line breaks and is read with
\code{rt_select_script_section(collapse=TRUE)}, \code{rt_has_argument} signals a failure.
}
\seealso{
\link{rt_select_script_section}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
