% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_class.R
\name{rt_has_class}
\alias{rt_has_class}
\title{test object class}
\usage{
rt_has_class(obj, class)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{class}{Charstring passed to \link{inherits}}
}
\value{
Logical: TRUE / FALSE depending on whether \code{obj} \link{inherits} the desired \link{class}.
}
\description{
Test class of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
