% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_column.R
\name{rt_has_column}
\alias{rt_has_column}
\title{test object column}
\usage{
rt_has_column(df, column, name = deparse(substitute(df)))
}
\arguments{
\item{df}{data.frame-like object to be tested.}

\item{column}{Charstring (vector) to which \link{colnames} is compared}

\item{name}{Object name for \link{rt_warn}. DEFAULT: object name}
}
\value{
Logical: TRUE / FALSE depending on whether object has the desired column(s).
}
\description{
Test colnames of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\code{names} and \code{ncols} in \link{rt_test_task}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
