% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_dim.R
\name{rt_has_dim}
\alias{rt_has_dim}
\title{check whether object has intended dimension}
\usage{
rt_has_dim(obj, value, name = deparse(substitute(obj)))
}
\arguments{
\item{obj}{Object to be checked}

\item{value}{Target object - NOT its dimension (for convenience).}

\item{name}{\link{rt_warn} name. DEFAULT: deparse(substitute(obj))}
}
\value{
TRUE or FALSE, depending on whether test is passed or failed
}
\description{
Checks \link{length} for vectors/lists, \link{nrow}+\link{ncol} for dataframe, matrix, array
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
