% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_length.R
\name{rt_has_length}
\alias{rt_has_length}
\title{test object dimension}
\usage{
rt_has_length(obj, len)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{len}{Value to which \link{length} is compared}
}
\value{
Logical: TRUE / FALSE depending on whether object has the desired length.
}
\description{
Test length of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
