% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_names.R
\name{rt_has_names}
\alias{rt_has_names}
\title{test object names}
\usage{
rt_has_names(obj, n)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{n}{Charstring to which \link{names} is compared}
}
\value{
Logical: TRUE / FALSE depending on whether object has the desired names.
}
\description{
Test names of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\link{rt_has_column}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
