% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_nrows.R
\name{rt_has_nrows}
\alias{rt_has_nrows}
\title{test object dimension}
\usage{
rt_has_nrows(df, len)
}
\arguments{
\item{df}{data.frame-like object to be tested. Just the name, not a character string.}

\item{len}{Value to which \link{nrow} is compared}
}
\value{
Logical: TRUE / FALSE depending on whether object has \code{len} rows.
}
\description{
Test number of rows of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\link{rt_has_ncols}, \link{rt_has_length}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
