% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_is_identical.R
\name{rt_is_identical}
\alias{rt_is_identical}
\title{test object identity}
\usage{
rt_is_identical(obj, target)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{target}{Object to be compared to via \link{identical}}
}
\value{
Logical: TRUE / FALSE depending on whether objects are identical.
}
\description{
Test whether two objects are identical, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
