% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_is_OS.R
\name{rt_is_OS}
\alias{rt_is_OS}
\title{Determine or check for OS platform}
\usage{
rt_is_OS(os = NULL)
}
\arguments{
\item{os}{Path to (processed) exercise folder.}
}
\value{
OS name if \code{os=NULL}, TRUE/FALSE otherwise (see examples)
}
\description{
Determine or check for OS platform.
Possible are: Linux, Windows, Darwin (or Mac), SunOS (for Solaris)
}
\examples{
rt_is_OS()
rt_is_OS("Linux")
rt_is_OS(c("windows","Linux")) # TRUE if one of those
rt_is_OS(c("Mac","Linux"))
rt_is_OS(c("Mac","windows"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2020
}
