% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_local_score.R
\name{rt_local_score}
\alias{rt_local_score}
\title{locally score complete exercise}
\usage{
rt_local_score(tfile = NULL)
}
\arguments{
\item{tfile}{Name of *tests.R file to be run.
Or name of exercise file, will be changed according to MOOC system.
DEFAULT: NULL (will be obtained from currently open File in Rstudio)}
}
\value{
Vector with number of total and passed tests, invisibly
}
\description{
Run test script for entire exercise without server access.
Intended for teacher / trainer use only.
Requires *tests.R script to be present (hidden on CodeOcean and not downloaded in zip folder).
}
\seealso{
\link{rt_score} for students. \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{internal}
\keyword{test}
