% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_add_opened_files.R
\name{rt_nextlargestid}
\alias{rt_nextlargestid}
\title{Find next largest 8 digit hexadecimal id}
\usage{
rt_nextlargestid(id)
}
\arguments{
\item{id}{Character string, see examples}
}
\value{
Hexadecimal ID charstring
}
\description{
Internal function to find the next largest 8 digit hexadecimal id
}
\examples{
\dontrun{ # non-exported function cannot have autorun examples
# random hex name:
id <- paste0(sample(c(0:9,LETTERS[1:6]), 8, TRUE), collapse="")
id; rt_nextlargestid(id)
rt_nextlargestid("ABCDEFAB")
rt_nextlargestid("ABCD9999")
rt_nextlargestid("ABCD6789")
rt_nextlargestid("12345678")
rt_nextlargestid("A0000000")
}

}
\seealso{
\code{\link[=rt_file2openedlist]{rt_file2openedlist()}}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{internal}
