% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_nextlargestorder.R
\name{rt_nextlargestorder}
\alias{rt_nextlargestorder}
\title{Find next largest relative order}
\usage{
rt_nextlargestorder(dir = ".")
}
\arguments{
\item{dir}{folder like ".Rproj.user/CONTEXT_ID/sources/per/t/"}
}
\value{
Hexadecimal ID charstring
}
\description{
Internal function to find the next largest relative order of scripts.
Only necessary on Mac OS, not Windows.
}
\seealso{
\code{\link[=rt_add_opened_files]{rt_add_opened_files()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2020
}
\keyword{internal}
