% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_no_globals.R
\name{rt_no_globals}
\alias{rt_no_globals}
\title{Check that a function does not use objects from the global environment}
\usage{
rt_no_globals(fun, name, globalok = NULL, qmark = TRUE)
}
\arguments{
\item{fun}{Function to check, as defined/sourced in the student script.}

\item{name}{Object name used in messages.}

\item{globalok}{Character vector of object names that are allowed to be
used as globals within \code{fun}. DEFAULT: NULL}

\item{qmark}{Include ' marks around \code{name}? DEFAULT: TRUE}
}
\value{
TRUE / FALSE
}
\description{
Detect whether a student-defined function references objects
defined at script level (globals) instead of using its own arguments/locals.
Default argument values such as \verb{function(x, y=default_val)} are stripped
before scanning, so they are never flagged (only the function body is).
}
\examples{
default_val <- 5
f1 <- function(x, y=default_val) x + y   # fine: default value, not flagged
rt_no_globals(f1, "f1")

f2 <- function(x) x + default_val        # uses a global inside the body
rt_no_globals(f2, "f2")

rt_no_globals(f2, "f2", globalok="default_val") # explicitly allowed

}
\seealso{
Used in \link{rt_test_object}, \link[codetools:findGlobals]{codetools::findGlobals}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2026
}
