% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_read_cofile.R
\name{rt_read_cofile}
\alias{rt_read_cofile}
\title{Read .co file}
\usage{
rt_read_cofile(cofile)
}
\arguments{
\item{cofile}{Path of .co file, may be relative.}
}
\value{
List with elements 'token', 'url', 'files'
}
\description{
Read .co file (from downloaded CO task).
}
\seealso{
\code{\link{rt_score}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2021
}
\keyword{file}
