% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_script.R
\name{rt_script_runs}
\alias{rt_script_runs}
\title{test whether script can be run}
\usage{
rt_script_runs(scriptobject)
}
\arguments{
\item{scriptobject}{Charstring output from \link{rt_run_script} or \link{rt_select_script_section}.
Will be checked with \code{!\link{isFALSE}()}.}
}
\value{
Logical. No \link{rt_warn} message issued, that already happens in \link{rt_run_script}
}
\description{
test whether script can be run
}
\examples{
# ToDo: link to file

}
\seealso{
\link{rt_run_script}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
