% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_select_script_section.R
\name{rt_select_script_section}
\alias{rt_select_script_section}
\title{Select section of a script}
\usage{
rt_select_script_section(scriptlines, task_nr, collapse = TRUE)
}
\arguments{
\item{scriptlines}{Charstring with several elements, normally output from \link{rt_run_script}}

\item{task_nr}{Task number to be found, eg for task_nr=3, the lines between \code{t3_start} and \code{t3_end}.}

\item{collapse}{Replace linebreaks with \verb{;} (e.g. for nice inclusion in messages)? DEFAULT: TRUE}
}
\value{
FALSE for failure, otherwise selected lines of the script,
by default collapsed (linebreaks replaced with ;) through \link{paste}
}
\description{
Select the section of the script code between \verb{t*_start} and \verb{t*_end} for further analysis in tests.
The code is not executed, as this is already done in \link{rt_run_script}.
}
\seealso{
\link{rt_run_script}, \link{rt_script_runs}, \link{rt_has_argument},
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
