% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_set_context_id.R
\name{rt_set_context_id}
\alias{rt_set_context_id}
\title{Try to set contextIdentifier from .Rproj.user ID}
\usage{
rt_set_context_id(dir = ".")
}
\arguments{
\item{dir}{Path to some folder containing a .Rproj file
and hence a (hidden) ".Rproj.user" folder. DEFAULT: "."}
}
\value{
Hexadecimal ID charstring
}
\description{
Try to find the User-specific 8 digit hexadecimal id from .Rproj.user ID.
Then save it to a file so that it can later be found by \code{\link[=rt_get_context_id]{rt_get_context_id()}}.
If successful, it will be saved in "~/.rstudio_context_id_by_codeoceanR.txt".
}
\seealso{
\code{\link[=rt_get_context_id]{rt_get_context_id()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2020
}
