% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_setup_exercise.R
\name{rt_setup_exercise}
\alias{rt_setup_exercise}
\title{Design complete exercise with tasks from database}
\usage{
rt_setup_exercise(df, dir = ".", folder = "Quiz1")
}
\arguments{
\item{df}{Data.frame with id, task, script. See example.}

\item{dir}{Directory, pasted together with \code{folder}. DEFAULT: "."}

\item{folder}{Folder at \code{dir}. DEFAULT: "Quiz1"}
}
\value{
dir from last call to \link{rt_add_task}
}
\description{
Set up an entire exercise (\code{script_1.R}, \code{script_1.R}, \code{...}, \code{tests.R}).
Uses the google spreadsheet as a database from which tasks can be selected by ID.
To request access to the spreadsheet, go to
\url{https://docs.google.com/spreadsheets/d/1ggSOYQ_veXgPmvA8cHCLnASkLUvc6-3t6UhJWPRPVoQ}
}
\examples{
if(FALSE) # suppress automated runs
rt_setup_exercise(read.table(header=TRUE, text="
id      task  script
Q11_1   1     1
Q10_2   2     1
Q12_10  3     2
Q11_2   4     3
Q11_3   5     3"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
