% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_task.R
\name{rt_signal_success}
\alias{rt_signal_success}
\title{Indicate a task succeeded}
\usage{
rt_signal_success(tnumber)
}
\arguments{
\item{tnumber}{Task number, preferably numeric.
Can likely also be a charstring, if all \code{\link[=rt_succeeded]{rt_succeeded()}} calls check for a character.}
}
\value{
TRUE, so that it an be used after the last "test" in a task test suite.
}
\description{
Add \code{tnumber} to the vector \code{taskenvironment$success} to indicate a task succeeded.
}
\seealso{
\link{rt_succeeded}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
