% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_submit.R
\name{rt_submit}
\alias{rt_submit}
\title{upload exercise to codeOcean}
\usage{
rt_submit(dir = ".")
}
\arguments{
\item{dir}{Path to (processed) exercise folder.
Must contain ".co", tests.R" and all the "script_n.R" files
referenced there, with the changes by the student.
DEFAULT: "."}
}
\value{
\code{\link[httr:content]{httr::content()}} output of response to http request, invisibly
}
\description{
upload complete exercise, message CO Score results
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
