% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_task.R
\name{rt_succeeded}
\alias{rt_succeeded}
\title{Test whether a (previous) task succeeded}
\usage{
rt_succeeded(tnumber, msg = "Please first solve task ")
}
\arguments{
\item{tnumber}{Task number, see \link{rt_signal_success}}

\item{msg}{Charsting with message to be prefixed to \code{tnumber}.}
}
\value{
TRUE or FALSE, depending on \code{taskenvironment$success} content.
}
\description{
Test whether \code{tnumber} is in the vector \code{taskenvironment$success}.
}
\seealso{
\link{rt_signal_success}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
