% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test.R
\name{rt_test}
\alias{rt_test}
\title{test condition}
\usage{
rt_test(condition, ...)
}
\arguments{
\item{condition}{Generic custom condition to be tested}

\item{\dots}{Message passed to \link{rt_warn} if condition evaluates to FALSE}
}
\value{
Logical: TRUE / FALSE depending on whether condition is TRUE.
}
\description{
Test a condition, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
