% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test_exercise.R
\name{rt_test_exercise}
\alias{rt_test_exercise}
\title{Wrapper for test script for complete exercise}
\usage{
rt_test_exercise(expr)
}
\arguments{
\item{expr}{calls to \link{rt_test_task}, in brackets}
}
\value{
NULL, writes to the console with \link{cat}.
}
\description{
Wrapper for test script for complete exercise
}
\examples{
rt_test_exercise(rnorm(20))

}
\seealso{
\link{rt_test_task}, tests.R in the \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
\keyword{test}
