% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_update_package.R
\name{rt_update_package}
\alias{rt_update_package}
\title{Update codeoceanR package to latest development version}
\usage{
rt_update_package(
  pack = "codeoceanR",
  user = "openHPI",
  quiet = FALSE,
  quietremotes = TRUE,
  ...
)
}
\arguments{
\item{pack}{Name of (already installed) package. DEFAULT: "codeoceanR "}

\item{user}{Github username. repo will then be user/pack. DEFAULT: "openHPI"}

\item{quiet}{Suppress version messages output?  DEFAULT: FALSE}

\item{quietremotes}{Suppress \code{remotes::install} output? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link[remotes:install_github]{remotes::install_github()}}}
}
\value{
data.frame with version information
}
\description{
Update codeoceanR to the latest development version on github, if necessary.
If the version number or date is larger on github,
\code{\link[remotes:install_github]{remotes::install_github()}} will be called.
}
\examples{
# rt_updatePackage()

}
\seealso{
\code{\link[remotes:install_github]{remotes::install_github()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2019 (rdwd), Oct 2020 (codeoceanR)
}
\keyword{file}
