% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_user_data_dir.R
\name{rt_user_data_dir}
\alias{rt_user_data_dir}
\title{Get directory for .Rproj.user folder}
\usage{
rt_user_data_dir()
}
\value{
path
}
\description{
Internal function to find the project-specific path for .Rproj.user
}
\seealso{
\code{\link[=rt_add_opened_files]{rt_add_opened_files()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2025
}
\keyword{internal}
