% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_warn.R
\name{rt_warn}
\alias{rt_warn}
\title{warn about errors}
\usage{
rt_warn(...)
}
\arguments{
\item{\dots}{Message components passed to \code{\link[=warning]{warning()}} or \code{\link[=cat]{cat()}}.
Can be named de or en to only be shown for a specific language,
see the examples}
}
\value{
FALSE and console printout of \code{\link[=message]{message()}} in interactive mode (RStudio) or
\code{\link[=cat]{cat()}} otherwise (CodeOcean)
}
\description{
pass test messages to Rstudio / CodeOcean.
}
\examples{
rt_warn("This is a CodeOcean message!")
rt_warn("This is message Nr ", 2+6, ".")
rt_env <- codeoceanR:::rt_env # not exported to keep UI clean.
rt_test_env <- list2env(rt_env(lang="en"))
rt_warn(de="de1 ", en="en1 ", 77, de=" de2", en=" en2", "'", min(longley))
invisible(rt_env(lang="de"))
rt_warn(de="de1 ", en="en1 ", 77, de=" de2", en=" en2", "'", min(longley))
# rt_env(lang="DE") # error message
rt_warn("Part1 ", if(TRUE) c(de="de3", en="en3"), " closing.") # both included.
cond <- TRUE
rt_warn("P1 ", de=if(cond)"de3 ", en=if(cond)"en3 ", "end.")
cond <- FALSE
rt_warn("P1 ", de=if(cond)"de3 ", en=if(cond)"en3 ", "end.")
rm(rt_test_env, rt_env)

}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{IO}
\keyword{error}
