% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_exes.R
\name{unzip_exes}
\alias{unzip_exes}
\title{Create exercise folders from zip files}
\usage{
unzip_exes(path = ".")
}
\arguments{
\item{path}{Exercise folder containing (several) downloaded zip files.
Use \code{path=NULL} to have it use \code{\link[rstudioapi:selectDirectory]{rstudioapi::selectDirectory()}}.
DEFAULT: "." (current working directory)}
}
\value{
exdir, invisibly
}
\description{
Create exercise folder from each existing zip file.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2026
}
\keyword{file}
