% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_local_score.R
\name{py_local_score}
\alias{py_local_score}
\title{locally score python exercise}
\usage{
py_local_score(tfile = NULL)
}
\arguments{
\item{tfile}{Name of *test.py file to be run.
DEFAULT: NULL (will be obtained from currently open File in Rstudio)}
}
\description{
Run python exercise test script without server access.
Intended for teacher / trainer use only.
Requires *test.py script to be present (hidden on CodeOcean and not downloaded in zip folder).
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2024
}
\keyword{test}
