% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_add_opened_files.R
\name{rt_add_opened_files}
\alias{rt_add_opened_files}
\title{Add files to list of Rstudio opened source documents}
\usage{
rt_add_opened_files(file, dir = ".", contextid = rt_get_context_id())
}
\arguments{
\item{file}{Filename to be added, without path, must be present at \code{dir}.}

\item{dir}{Target directory with a .Rproj file. DEFAULT: "." \code{\link[=getwd]{getwd()}}}

\item{contextid}{Hexadecimal user ID, see \code{\link[=rt_get_context_id]{rt_get_context_id()}}}
}
\value{
expanded filename, invisibly.
}
\description{
Add files to the list of Rstudio opened source documents.
The related .Rproj Rstudio project should be closed at the time of writing.
}
\seealso{
\code{\link[=rt_create_task]{rt_create_task()}}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{internal}
