% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_contains.R
\name{rt_contains}
\alias{rt_contains}
\title{Test whether object contains certain value(s)}
\usage{
rt_contains(
  object,
  value,
  msgval = value,
  name = deparse(substitute(object)),
  qmark = TRUE,
  fixed = TRUE,
  remcom = NA,
  ignore_space = TRUE,
  ignore_quote = TRUE
)
}
\arguments{
\item{object}{Object to be tested. Just the name, not a character string.}

\item{value}{value(s) that should be in \code{object}, can be char / other.}

\item{msgval}{Value(s) to be messaged in the warning. Must be same length as value!
Enables testing with regex but displaying a wanted solution.
Remember to set fixed=FALSE as well. DEFAULT: \code{value}}

\item{name}{\link{rt_warn} name. DEFAULT: deparse(substitute(obj))}

\item{qmark}{Include ' marks around \code{name}? DEFAULT: TRUE}

\item{fixed}{Fixed match in \link{grepl}? DEFAULT: TRUE}

\item{remcom}{Remove comments from script before checking presence?
Experimental feature, uses \code{\link[=parse]{parse()}}.
DEFAULT: NA (yes, if object has length >1  and \code{value}
is a character string)}

\item{ignore_space}{Remove spaces before comparison? DEFAULT: TRUE}

\item{ignore_quote}{Replace \verb{'} with \verb{"} before comparison? DEFAULT: TRUE}
}
\value{
Logical: TRUE / FALSE depending on whether \code{object} contains \code{value}.
}
\description{
Test whether object contains certain value(s)
}
\examples{
rt_contains(1:7, 6)
rt_contains(1:7, 8)
rt_contains(1:7, 5:6)
rt_contains(1:7, 5:9)
stri <- "some char string"
rt_contains(stri, "arstring")
rt_contains(stri, "arstring", ignore_space=FALSE)
rt_contains(stri, c("some", "char", "STRING"))

}
\seealso{
\link{rt_has_value}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020 + 2021
}
\keyword{error}
\keyword{test}
