% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_create_task.R
\name{rt_create_task}
\alias{rt_create_task}
\title{Create Task folder from zip file}
\usage{
rt_create_task(
  zipfile = file.choose(),
  exdir = NULL,
  isunzipped = rt_is_OS("Windows"),
  deletezip = TRUE,
  ...
)
}
\arguments{
\item{zipfile}{Path to zip file. Remember: on Windows, "\" must be changed to "/".
DEFAULT: \code{\link[=file.choose]{file.choose()}}}

\item{exdir}{Folder to unzip to, e.g. "./task" for folder at current wd.
\code{exdir} may not yet exist, to avoid overwriting previously
unzipped and potentially edited tasks.
\code{exdir} must remain NULL if \code{isunzipped=TRUE}.
DEFAULT: NULL (path from zipfile).}

\item{isunzipped}{Is the quiz already unzipped to a folder? Then that folder will be used.
DEFAULT: \link{rt_is_OS}\code{("Mac")}}

\item{deletezip}{If the task was created sucessfully, delete the original zip file? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{unzip}}}
}
\value{
exdir, invisibly
}
\description{
Create task folder from zip file, with a \code{.Rproj} file and try to open that in Rstudio.
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
