% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_gives_echo.R
\name{rt_gives_echo}
\alias{rt_gives_echo}
\title{Does running code produce a wanted echo (\link{message} + \link{cat} + \link{print} outputs)?}
\usage{
rt_gives_echo(
  expr,
  e = NULL,
  value = FALSE,
  echo = FALSE,
  fixed = TRUE,
  exact = FALSE,
  msg = deparse(substitute(expr))
)
}
\arguments{
\item{expr}{Code to be executed while output is collected.
Can be contained within \{curly brackets\}.}

\item{e}{Charstring: echo that needs to be present (matched with \link{grepl}).
Use \code{e=""} to test if no echo is generated.
Use \code{e=NULL} to test if any echo is produced.
DEFAULT: NULL}

\item{value}{Return the result of running \code{expr}? DEFAULT: FALSE}

\item{echo}{Return the echo of running \code{expr} (instead of testing it)? DEFAULT: FALSE}

\item{fixed}{If tested for specific e with \link{grepl}, treat charstring \code{e} as is,
rather than regex? DEFAULT: TRUE}

\item{exact}{use == instead of grepl? Useful to see if message is used, not print.
DEFAULT: FALSE}

\item{msg}{Charstring to be included in testing messages. DEFAULT: input code}
}
\value{
TRUE / FALSE, Charstring vector from \code{\link[=readLines]{readLines()}}
}
\description{
Does running code produce a wanted echo (\link{message} + \link{cat} + \link{print} outputs)?
}
\examples{
rt_test_env <- new.env()
rt_gives_echo(cat("things\nwith linebreaks"), value=TRUE)
rt_gives_echo(message("some message"), value=TRUE)
rt_gives_echo({77; 88; 99}, value=TRUE) # empty
rt_gives_echo({77; print(88); 99}, value=TRUE) # "[1] 88"
rt_gives_echo({print(77); log(-6); print(99)}, value=TRUE) # warning not in echo
rt_gives_echo({print(77); log("6"); print(99)}, value=TRUE) # error is included

rt_gives_echo({print(77); message("some stuff")}, "e stuff") # TRUE
rt_gives_echo({print(77); message("some stuff")}, "more stuff") # wrong echo
rt_gives_echo(log(7), "") # TRUE (no echo)
rt_gives_echo(cat(88), "") # shouldn't echo but does
rt_gives_echo(cat(999), NULL) # TRUE (generates any echo)
rt_gives_echo(log(7), NULL) # should echo something but doesn't

}
\seealso{
\link{rt_gives_warning} \code{\link[=capture.output]{capture.output()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2020 + Oct 2021
}
