% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_args.R
\name{rt_has_args}
\alias{rt_has_args}
\title{Check code chunk with function call for correct arguments, e.g. in plotting tasks.}
\usage{
rt_has_args(
  code,
  target,
  snumber,
  nameonly = FALSE,
  stepwise = NULL,
  alt = NULL,
  opt = NULL,
  ignAssign = FALSE
)
}
\arguments{
\item{code}{Charstring with user (student) code.}

\item{target}{Charstring with desired code.}

\item{snumber}{Section number for messaging.}

\item{nameonly}{Check for e.g. 1:5 exactly, disallowing c(1,2,3,4,5).
DEFAULT: FALSE}

\item{stepwise}{Passed to \link{rt_test_object}. DEFAULT: NULL}

\item{alt}{list of alternately accepted inputs.
Use list(argname="anyval") to skip value test. DEFAULT: NULL}

\item{opt}{Charstring vector of arguments that are optional.
Mostly for the formula interface, I guess. DEFAULT: NULL}

\item{ignAssign}{Remove assignment part from code? DEFAULT: FALSE}
}
\value{
TRUE / FALSE
}
\description{
Check student code for certain argument values. Matching is done internally.
}
\details{
Used to be covered by \link{rt_has_argument}, which might be deprecated.
}
\examples{
rt_has_args("plot(1:5, lwd=2)", "plot(1:5, lwd=3)", 7)
# Tx: code section t7: argument 'lwd' should be '3', not '2'.

# 'data' arguments are attached internally, hence this works:
rt_has_args("plot(rate~conc,data=Puromycin,col=state)",
            "plot(rate~conc,data=Puromycin,col=state)", 7)

rt_has_args("plot(1:5, lwd=2)", "plot(1:5, lwd=3)", 7, alt=list(lwd=2:4)) # TRUE

rt_has_args("plot(c(1,2,3,4,5))", "plot(1:5)", 7               ) # TRUE
rt_has_args("plot(c(1,2,3,4,5))", "plot(1:5)", 7, nameonly=TRUE) # FALSE:
# Tx: code section t7: argument 'x' should be '1:5', not 'c(1, 2, 3, 4, 5)'.

# See more in unit tests at "codeoceanR/tests/testthat/argtests.R"

}
\seealso{
\link{str2lang}, \link{match.call}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2021
}
