% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_class.R
\name{rt_has_class}
\alias{rt_has_class}
\title{Test class of object}
\usage{
rt_has_class(object, class, name = deparse(substitute(object)))
}
\arguments{
\item{object}{Object. Regular object name, not quoted.}

\item{class}{Class(es) that are acceptable. Test is passed if any of the classes matches.
Test is skipped if \code{class="any"}.}

\item{name}{\link{rt_warn} name. DEFAULT: deparse(substitute(obj))}
}
\value{
TRUE / FALSE
}
\description{
Test class of object
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
