% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_value.R
\name{rt_has_value}
\alias{rt_has_value}
\title{test object value}
\usage{
rt_has_value(
  obj,
  value,
  digits = 6,
  name = deparse(substitute(obj)),
  noise = FALSE,
  stepwise = length(obj) > 1
)
}
\arguments{
\item{obj}{Object to be tested. Just the name, not a character string.}

\item{value}{value \code{obj} should have, can be char / numeric / other.}

\item{digits}{Tolerance - both \code{obj} and \code{value} are \link{round}ed before comparison. DEFAULT: 6}

\item{name}{Object name for \link{rt_warn} messages. DEFAULT: \code{deparse(substitute(obj))}}

\item{noise}{Add noise, so not the exact difference is reported? DEFAULT: FALSE}

\item{stepwise}{Compare vectors (in future also: matrices and data.frames)
stepwise? only if noise=FALSE. DEFAULT: length(obj)>1}
}
\value{
Logical: TRUE / FALSE depending on whether object has the given value.
}
\description{
Test values of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\link{rt_contains}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
