% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_has_value.R
\name{rt_has_value}
\alias{rt_has_value}
\title{test object value}
\usage{
rt_has_value(
  object,
  value,
  name = deparse(substitute(object)),
  qmark = TRUE,
  digits = 6,
  noise = FALSE,
  stepwise = NULL
)
}
\arguments{
\item{object}{Object to be tested. Just the name, not a character string.}

\item{value}{value \code{object} should have, can be char / numeric / other.}

\item{name}{Object name for \link{rt_warn} messages. DEFAULT: \code{deparse(substitute(object))}}

\item{qmark}{Include ' marks around \code{name}? DEFAULT: TRUE}

\item{digits}{Tolerance - both \code{object} and \code{value} are \link{round}ed before comparison,
if target \code{value} is numeric. DEFAULT: 6}

\item{noise}{Add noise, so not the exact difference is reported?
Only used if \code{value} is numeric. DEFAULT: FALSE}

\item{stepwise}{Compare \code{object} and \code{value} stepwise?
For numerical objects: only used if noise=FALSE.
DEFAULT: NULL, meaning \code{length(value)>1}}
}
\value{
Logical: TRUE / FALSE depending on whether object has the given value.
}
\description{
Test values of object, calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful message if needed.
}
\seealso{
\link{rt_contains}, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
