% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_local_score.R
\name{rt_local_score}
\alias{rt_local_score}
\title{locally score complete exercise}
\usage{
rt_local_score(dir = ".", check_unsaved = TRUE)
}
\arguments{
\item{dir}{Path to exercise folder.
Must contain "tests.R" and all the "script_n.R" files referenced there.
DEFAULT: "."}

\item{check_unsaved}{Check if any files are unsaved? Requires dir to be in an .Rproj.
Runs \code{\link[=rt_check_for_unsaved_files]{rt_check_for_unsaved_files()}}. DEFAULT: TRUE}
}
\value{
Vector with number of total and passed tests, invisibly
}
\description{
Run test script for entire exercise without server access.
Intended for teacher / trainer use only.
Requires tests.R script to be present (hidden on CodeOcean and not downloaded in zip folder).
}
\seealso{
\link{rt_score} for students. \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
