% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot.R
\name{rt_plot1}
\alias{rt_plot1}
\alias{rt_plot2}
\title{handle graphics on CodeOcean}
\usage{
rt_plot1(...)

rt_plot2(maxn = 10)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[=png]{png()}}}

\item{maxn}{Maximum number of figures to plot. Note: rendering time increases drastically!}
}
\description{
Display graphics in the CodeOcean Browser instance through base64 encoding.
}
\examples{
# Sys.setenv(CODEOCEAN="true")
rt_plot1() # for graphics in CodeOcean Browser instance
plot(1:10)
77 + 88
plot(rnorm(200))
rt_plot2() # preceding plots are displayed at the end of CO printout

}
\seealso{
\code{\link[=png]{png()}}, \code{\link[base64enc:base64]{base64enc::base64encode()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2022
}
\keyword{file}
