% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_run_script.R
\name{rt_run_script}
\alias{rt_run_script}
\title{run student-submitted scripts}
\usage{
rt_run_script(filename, quiet = TRUE)
}
\arguments{
\item{filename}{Path of script to be run}

\item{quiet}{If TRUE, suppress printed output, messages and warnings.
Errors are handled separately. DEFAULT: TRUE}
}
\value{
FALSE for failure, otherwise contents of the script, output of \code{\link[=readLines]{readLines()}}
}
\description{
Run script, with informative \link{rt_warn} message if \code{\link[=source]{source()}} fails.
}
\seealso{
\link{rt_test_task}, \link{rt_select_script_section},
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
