% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_score.R
\name{rt_score}
\alias{rt_score}
\title{score complete exercise}
\usage{
rt_score(dir = ".")
}
\arguments{
\item{dir}{Path to exercise folder.
Must contain "tests.R" and all the "script_n.R" files referenced there.
DEFAULT: "."}
}
\value{
Vector with number of total and passed tests, invisibly
}
\description{
Run test script for entire exercise
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
