% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_score.R
\name{rt_score}
\alias{rt_score}
\title{upload exercise to codeOcean}
\usage{
rt_score(dir = ".", final = FALSE, wait = 15, submit = FALSE)
}
\arguments{
\item{dir}{Path to (processed) exercise folder.
Must contain ".co" and all the "script_n.R" files
referenced there, with the changes by the student, saved.
DEFAULT: "."}

\item{final}{Print instructions (+ open URL, if available) for final submission?
DEFAULT: FALSE}

\item{wait}{Number of seconds to wait before opening the url.
If >60 seconds, it is set to 60 seconds internally. DEFAULT: 15}

\item{submit}{Submit grade to openHPI? To be called from \code{\link[=rt_submit]{rt_submit()}}.
Will soon replace the 'final' option}
}
\value{
\code{\link[httr:content]{httr::content()}} output of response to http request, invisibly
}
\description{
upload complete exercise, message CO Score results
}
\seealso{
\link{rt_local_score} for teachers, \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
