% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_score.R
\name{rt_score}
\alias{rt_score}
\title{score complete exercise}
\usage{
rt_score(tfile = "tests.R")
}
\arguments{
\item{tfile}{Path to tests.R file. ToDo: automated detection. DEFAULT: "tests.R"}
}
\value{
Vector with number of total and passed tests, invisibly
}
\description{
Run test script for entire exercise
}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{test}
