% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_script_section.R
\name{rt_script_section}
\alias{rt_script_section}
\title{Select section of a script}
\usage{
rt_script_section(
  script,
  snumber,
  name = deparse(substitute(script)),
  collapse = NULL,
  maxlen = 95
)
}
\arguments{
\item{script}{Charstring with several elements, normally output from \link{rt_run_script}.}

\item{snumber}{Section to be found, eg for snumber=3, the lines between \code{t3_start} and \code{t3_end}.}

\item{name}{\link{rt_warn} name. DEFAULT: deparse(substitute(script))}

\item{collapse}{Replacement for linebreaks. For nice inclusion in messages or
custom grepl tests without if(grepl(...)) error the condition has length > 1.
DEFAULT: NULL (not collapsed)}

\item{maxlen}{Maximum allowed length of lines of code. DEFAULT: 95}
}
\value{
FALSE for failure, otherwise selected lines of the script,
by default NOT collapsed (linebreaks replaced with ;) through \link{paste}
}
\description{
Select the section of the script code between \verb{t*_start} and \verb{t*_end} for further analysis in tests.
}
\seealso{
\link{rt_test_task}, \link{rt_run_script}, \link{rt_has_args},
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020 + Nov 2021
}
\keyword{file}
