% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_select_script_section.R
\name{rt_select_script_section}
\alias{rt_select_script_section}
\title{Select section of a script}
\usage{
rt_select_script_section(
  scriptlines,
  task_nr,
  msg_nr = task_nr,
  name = deparse(substitute(scriptlines)),
  collapse = FALSE,
  maxlen = 95
)
}
\arguments{
\item{scriptlines}{Charstring with several elements, normally output from \link{rt_run_script}}

\item{task_nr}{Task number to be found, eg for task_nr=3, the lines between \code{t3_start} and \code{t3_end}.}

\item{msg_nr}{Task number for \link{rt_warn}. DEFAULT: task_nr}

\item{name}{\link{rt_warn} name. DEFAULT: deparse(substitute(scriptlines))}

\item{collapse}{Replace linebreaks with \verb{;} (e.g. for nice inclusion in messages)?
Also helpful if you want do run your own grepl tests on it
to not have a vector and if(grepl(...)) error the condition has length > 1.
DEFAULT: FALSE}

\item{maxlen}{Maximum length of lines of code. DEFAULT: 95}
}
\value{
FALSE for failure, otherwise selected lines of the script,
by default NOT collapsed (linebreaks replaced with ;) through \link{paste}
}
\description{
Select the section of the script code between \verb{t*_start} and \verb{t*_end} for further analysis in tests.
The code is not executed, as this is already done in \link{rt_run_script}.\\
Attenton: if the user script contains line breaks and is read with
\code{collapse=TRUE}, \link{rt_has_argument} leads to a test script failure!
}
\seealso{
\link{rt_test_task}, \link{rt_run_script}, \link{rt_has_argument},
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{file}
