% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test_object.R
\name{rt_test_object}
\alias{rt_test_object}
\title{check whether object has intended class, dimension, names and value}
\usage{
rt_test_object(
  object,
  value,
  name = deparse(substitute(object)),
  qmark = TRUE,
  class = NULL,
  intnum = TRUE,
  dim = TRUE,
  names = TRUE,
  hasval = TRUE,
  stepwise = NULL,
  stepnames = FALSE
)
}
\arguments{
\item{object, value}{object to be checked against value}

\item{name}{Object name used in messages}

\item{qmark}{Include ' marks around \code{name}? DEFAULT: TRUE}

\item{class, intnum, dim, names, hasval, stepwise, stepnames}{See \link{rt_test_task}}
}
\value{
TRUE or FALSE, depending on whether full test is passed or failed
}
\description{
check whether object has intended class, dimension, names and value
}
\seealso{
Used in \link{rt_test_task}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
