% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test_object.R
\name{rt_test_object}
\alias{rt_test_object}
\alias{rt_exists}
\alias{rt_has_class}
\alias{rt_has_length}
\alias{rt_has_ncols}
\alias{rt_has_nrows}
\alias{rt_has_column}
\alias{rt_has_names}
\alias{rt_is_identical}
\alias{rt_contains}
\alias{rt_has_value}
\title{test objects in CodeOcean Setting}
\usage{
rt_exists(obj)

rt_has_class(obj, class)

rt_has_length(obj, len)

rt_has_ncols(df, len)

rt_has_nrows(df, len)

rt_has_column(df, column)

rt_has_names(obj, n)

rt_is_identical(obj, target)

rt_contains(obj, value)

rt_has_value(obj, value, digits = 6, noise = TRUE)
}
\arguments{
\item{obj}{Object to be tested}

\item{class}{Charstring passed to \link{inherits}}

\item{len}{Value to which \link{length}, \link{ncol} or \link{nrow} is compared}

\item{df}{Object (data.frame) to be tested}

\item{column}{Charstring to which \link{colnames} is compared}

\item{n}{Charstring to which \link{names} is compared}

\item{target}{Object to be compared to via \link{identical}}

\item{value}{value \code{obj} should have, can be char / numeric / other.}

\item{digits}{Tolerance - both \code{obj} and \code{value} are \link{round}ed before comparison. DEFAULT: 6}

\item{noise}{Add noise, so not the exact difference is reported? DEFAULT: TRUE}
}
\value{
Logical: TRUE / FALSE depending on whether condition is met
}
\description{
Diverse object tests calling \code{\link[=rt_warn]{rt_warn()}} with informative and helpful messages.
}
\examples{
# ToDo / link to example test file

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{error}
\keyword{test}
