% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test_exercise.R
\name{rt_test_task}
\alias{rt_test_task}
\title{Wrapper for test script for an individual task within an exercise}
\usage{
rt_test_task(
  tnumber,
  ...,
  script = NULL,
  object = NULL,
  class = NULL,
  length = NULL,
  nrows = NULL,
  ncols = NULL,
  names = NULL,
  value = NULL,
  noise = FALSE
)
}
\arguments{
\item{tnumber}{Number of task.}

\item{\dots}{Further tests, comma-separated}

\item{script}{DEFAULT: NULL}

\item{object}{DEFAULT: NULL}

\item{class}{DEFAULT: NULL}

\item{length}{DEFAULT: NULL}

\item{nrows}{DEFAULT: NULL}

\item{ncols}{DEFAULT: NULL}

\item{names}{DEFAULT: NULL}

\item{value}{DEFAULT: NULL}

\item{noise}{DEFAULT: FALSE}
}
\value{
TRUE or FALSE, indicating whterh all tests passed
}
\description{
Wrapper for test script for an individual task within an exercise
}
\seealso{
\link{rt_test_exercise}, tests.R in the \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
