% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_test_task.R
\name{rt_test_task}
\alias{rt_test_task}
\title{Wrapper for test script for an individual task within an exercise}
\usage{
rt_test_task(
  tnumber,
  script = NULL,
  object = NULL,
  value = NULL,
  ...,
  zero = hasval,
  class = NULL,
  hasval = TRUE,
  dim = hasval,
  correct = !hasval,
  df_test = TRUE,
  noise = FALSE,
  stepwise = NULL,
  solved = NULL,
  inputs = NULL,
  export = NULL,
  names = FALSE
)
}
\arguments{
\item{tnumber}{Number of task. Must be numeric, as it will be used for pass/fail in \link{rt_env}.}

\item{script}{Exercise script content that needs to be non-empty.
This is the only test that will not generate an \link{rt_warn} message,
as \link{rt_run_script} or \link{rt_select_script_section} already do.
DEFAULT: NULL}

\item{object}{Object that needs to be created in student script.
Regular object name, not quoted. DEFAULT: NULL}

\item{value}{Intended value. Existence, class and dimensions are checked first,
then \link{rt_has_value} is called if hasval=TRUE. DEFAULT: NULL}

\item{\dots}{Further tests, comma-separated, can be given at the end.
Need to yield TRUE/FALSE and take care of \link{rt_warn} individually.
The \verb{rt_*} test functions are designed for just that :).}

\item{zero}{Check for pre-assigned objects (to 0) with special message? DEFAULT: hasval}

\item{class}{Class(es) that are acceptable. Test is passed if any of the classes matches.
Test is skipped if \code{class="any"}. DEFAULT: NULL (class of \code{value})}

\item{hasval}{After checks for existence, class and dimension, run \link{rt_has_value}?
Can be FALSE for custom messages, e.g in multiple choice tasks.
Will be FALSE if \code{value} is a function.
DEFAULT: TRUE}

\item{dim}{Check dimension (length or nrow+ncol) with \link{rt_has_dim}?
Will be FALSE if \code{value} is a function. DEFAULT: hasval}

\item{correct}{Custom value message for multiple choice tasks?
Will be FALSE if \code{value} is a function. DEFAULT: !hasval}

\item{df_test}{Run tests on data.frames? Set to FALSE if \code{!is.data.frame(value)}.
Checks \link{rt_has_column}, classes and then \link{rt_has_value} per column.
DEAULT: TRUE}

\item{noise}{noise parameter in \link{rt_has_value}. DEFAULT: FALSE}

\item{stepwise}{stepwise parameter in \link{rt_has_value}. DEFAULT: NULL}

\item{solved}{Task number that must be solved before other tests are run. DEFAULT: NULL}

\item{inputs}{List or vector with (named) charstrings with code to be called
if \code{object} and \code{value} are functions.
Will be called with \code{eval(str2lang(paste0("object(",input_i,")")))}.
Object names within rt_test_exercise are not available within rt_test_task,
unlisst listed in \code{export}.
For single-argument functions, numerical input works fine, too.
These are the only tests run AFTER ...-tests are run.
DEFAULT: NULL}

\item{export}{Character vector with object names to be assigned into evaluation environment.
DEFAULT: NULL}

\item{names}{Test whether \code{object} has the same \link{names} as \code{value}? DEFAULT: FALSE}
}
\value{
TRUE or FALSE, indicating whether all tests passed
}
\description{
Wrapper for test script for an individual task within an exercise
}
\details{
NULL arguments mean that their respective tests are not performed.\cr
Tests are performed in an order for increasingly specific
messages that do not give away the solution too early.\cr
Arguments after \dots need to be specified by full name.
\code{tnumber} must always be given, \code{script}, \code{object} and \code{value}
must be given (can be NULL) so later custom tests are not matched to them.
}
\seealso{
\link{rt_test_exercise}, tests.R in the \href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
