% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_vec.R
\name{rt_vec}
\alias{rt_vec}
\title{Convert vectors to charstring code representations}
\usage{
rt_vec(x, n = NULL, l = -99, u = 999)
}
\arguments{
\item{x}{input vector}

\item{n}{number of integers to sample between \code{l} and \code{u}. DEFAULT: NULL}

\item{l, u}{Lower and upper boundaries in sample, if n is given.}
}
\value{
single charstring
}
\description{
Convert vectors to charstring code representations
}
\examples{
rt_vec(pi)
rt_vec("hey")
message(rt_vec("hey"))
message(rt_vec(letters[1:6]))
message(rt_vec(sample(1:90, 7)))
message(rt_vec(n=5))
}
\seealso{
\link{dput}, \link{toString}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2021
}
\keyword{character}
