% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_warn.R
\name{rt_warn}
\alias{rt_warn}
\title{warn about errors}
\usage{
rt_warn(...)
}
\arguments{
\item{\dots}{Message components passed to \code{\link[=warning]{warning()}} or \code{\link[=cat]{cat()}}.}
}
\value{
invisible FALSE and console printout of \code{\link[=message]{message()}} in interactive mode (RStudio) or
\code{\link[=cat]{cat()}} otherwise (CodeOcean)
}
\description{
pass test messages to Rstudio / CodeOcean.
}
\examples{
rt_warn("This is a CodeOcean message!")

}
\seealso{
\href{https://github.com/openHPI/codeoceanR/tree/main/inst/extdata}{exercise example} on github
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2020
}
\keyword{IO}
\keyword{error}
